# articurated/API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: ArtiCurated Order & Returns Management API
  description: Backend API for managing the full lifecycle of orders and returns
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development

tags:
  - name: Orders
    description: Order lifecycle management
  - name: Returns
    description: Return request lifecycle management

paths:
  /orders:
    post:
      tags: [Orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              customerName: "Alice Smith"
              customerEmail: "alice@example.com"
              items:
                - productName: "Handcrafted Vase"
                  productSku: "VASE-001"
                  quantity: 2
                  unitPrice: 49.99
      responses:
        '201':
          description: Order created in PENDING_PAYMENT state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [Orders]
      summary: List all orders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /orders/{orderId}/status:
    patch:
      tags: [Orders]
      summary: Transition order status
      description: |
        Valid transitions:
        - PENDING_PAYMENT → PAID
        - PENDING_PAYMENT → CANCELLED
        - PAID → PROCESSING_IN_WAREHOUSE
        - PAID → CANCELLED
        - PROCESSING_IN_WAREHOUSE → SHIPPED  *(triggers PDF invoice background job)*
        - SHIPPED → DELIVERED
      parameters:
        - $ref: '#/components/parameters/orderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderTransitionRequest'
            example:
              targetStatus: "PAID"
              notes: "Payment confirmed via Stripe"
              changedBy: "payment-service"
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '409':
          $ref: '#/components/responses/InvalidTransition'
        '404':
          $ref: '#/components/responses/NotFound'

  /orders/{orderId}/history:
    get:
      tags: [Orders]
      summary: Get order status audit trail
      parameters:
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Status history records ordered by time ascending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusHistoryResponse'

  /orders/{orderId}/returns:
    post:
      tags: [Returns]
      summary: Initiate a return request
      description: Order must be in DELIVERED state
      parameters:
        - $ref: '#/components/parameters/orderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReturnRequest'
            example:
              reason: "Item arrived with a crack on the base"
      responses:
        '201':
          description: Return created in REQUESTED state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnResponse'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'

    get:
      tags: [Returns]
      summary: List all returns for an order
      parameters:
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: List of return requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReturnResponse'

  /returns/{returnId}:
    get:
      tags: [Returns]
      summary: Get return request by ID
      parameters:
        - $ref: '#/components/parameters/returnId'
      responses:
        '200':
          description: Return request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /returns/{returnId}/status:
    patch:
      tags: [Returns]
      summary: Transition return status
      description: |
        Valid transitions:
        - REQUESTED → APPROVED (manager action)
        - REQUESTED → REJECTED (manager action)
        - APPROVED → IN_TRANSIT (customer ships item back)
        - IN_TRANSIT → RECEIVED (warehouse confirms receipt)
        - RECEIVED → COMPLETED *(triggers refund background job)*
      parameters:
        - $ref: '#/components/parameters/returnId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnTransitionRequest'
            example:
              targetStatus: "APPROVED"
              notes: "Item verified as damaged in photos"
              changedBy: "manager-jane"
      responses:
        '200':
          description: Updated return request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnResponse'
        '409':
          $ref: '#/components/responses/InvalidTransition'

  /returns/{returnId}/history:
    get:
      tags: [Returns]
      summary: Get return status audit trail
      parameters:
        - $ref: '#/components/parameters/returnId'
      responses:
        '200':
          description: Status history records ordered by time ascending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusHistoryResponse'

components:
  parameters:
    orderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    returnId:
      name: returnId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    CreateOrderRequest:
      type: object
      required: [customerName, customerEmail, items]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemRequest:
      type: object
      required: [productName, productSku, quantity, unitPrice]
      properties:
        productName:
          type: string
        productSku:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: decimal
          minimum: 0.01

    OrderTransitionRequest:
      type: object
      required: [targetStatus]
      properties:
        targetStatus:
          type: string
          enum: [PAID, PROCESSING_IN_WAREHOUSE, SHIPPED, DELIVERED, CANCELLED]
        notes:
          type: string
        changedBy:
          type: string

    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [PENDING_PAYMENT, PAID, PROCESSING_IN_WAREHOUSE, SHIPPED, DELIVERED, CANCELLED]
        invoicePath:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    OrderItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productName:
          type: string
        productSku:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number

    CreateReturnRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    ReturnTransitionRequest:
      type: object
      required: [targetStatus]
      properties:
        targetStatus:
          type: string
          enum: [APPROVED, REJECTED, IN_TRANSIT, RECEIVED, COMPLETED]
        notes:
          type: string
        changedBy:
          type: string
        trackingNumber:
          type: string
          description: Required when transitioning to IN_TRANSIT

    ReturnResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        orderNumber:
          type: string
        status:
          type: string
          enum: [REQUESTED, APPROVED, REJECTED, IN_TRANSIT, RECEIVED, COMPLETED]
        reason:
          type: string
        reviewNotes:
          type: string
          nullable: true
        trackingNumber:
          type: string
          nullable: true
        refundTransactionId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    StatusHistoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fromStatus:
          type: string
        toStatus:
          type: string
        notes:
          type: string
          nullable: true
        changedBy:
          type: string
        changedAt:
          type: string
          format: date-time

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              status: {type: integer, example: 404}
              message: {type: string}
              timestamp: {type: string, format: date-time}

    InvalidTransition:
      description: Invalid state transition
      content:
        application/json:
          schema:
            type: object
            properties:
              status: {type: integer, example: 409}
              message: {type: string}
              timestamp: {type: string, format: date-time}

    BusinessRuleViolation:
      description: Business rule violation
      content:
        application/json:
          schema:
            type: object
            properties:
              status: {type: integer, example: 422}
              message: {type: string}
              timestamp: {type: string, format: date-time}

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              status: {type: integer, example: 400}
              message: {type: string}
              errors: {type: object}
              timestamp: {type: string, format: date-time}